history: updated statistics history_id to bigint
Bug-Url: https://bugzilla.redhat.com/1121994

+++ etl_export/src/ovirt_engine_dwh/aggregationtodaily_3_5/AggregationToDaily.java
@@ -1532,9 +1532,9 @@ public class AggregationToDaily implements TalendJob {
 		final static byte[] commonByteArrayLock_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 		static byte[] commonByteArray_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 
-		public int history_id;
+		public long history_id;
 
-		public int getHistory_id() {
+		public long getHistory_id() {
 			return this.history_id;
 		}
 
@@ -1599,7 +1599,7 @@ public class AggregationToDaily implements TalendJob {
 
 					int length = 0;
 
-					this.history_id = dis.readInt();
+					this.history_id = dis.readLong();
 
 					this.history_datetime = readDate(dis);
 
@@ -1626,9 +1626,9 @@ public class AggregationToDaily implements TalendJob {
 		public void writeData(ObjectOutputStream dos) {
 			try {
 
-				// int
+				// long
 
-				dos.writeInt(this.history_id);
+				dos.writeLong(this.history_id);
 
 				// java.util.Date
 
@@ -2121,7 +2121,7 @@ public class AggregationToDaily implements TalendJob {
 							if (rs_tJDBCInput_1
 									.getObject(column_index_tJDBCInput_1) != null) {
 								row1.history_id = rs_tJDBCInput_1
-										.getInt(column_index_tJDBCInput_1);
+										.getLong(column_index_tJDBCInput_1);
 							} else {
 
 								throw new RuntimeException(
@@ -4381,9 +4381,9 @@ public class AggregationToDaily implements TalendJob {
 		final static byte[] commonByteArrayLock_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 		static byte[] commonByteArray_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 
-		public int history_id;
+		public long history_id;
 
-		public int getHistory_id() {
+		public long getHistory_id() {
 			return this.history_id;
 		}
 
@@ -4536,7 +4536,7 @@ public class AggregationToDaily implements TalendJob {
 
 					int length = 0;
 
-					this.history_id = dis.readInt();
+					this.history_id = dis.readLong();
 
 					this.history_datetime = readDate(dis);
 
@@ -4625,9 +4625,9 @@ public class AggregationToDaily implements TalendJob {
 		public void writeData(ObjectOutputStream dos) {
 			try {
 
-				// int
+				// long
 
-				dos.writeInt(this.history_id);
+				dos.writeLong(this.history_id);
 
 				// java.util.Date
 
@@ -5261,7 +5261,7 @@ public class AggregationToDaily implements TalendJob {
 							if (rs_tJDBCInput_2
 									.getObject(column_index_tJDBCInput_2) != null) {
 								row2.history_id = rs_tJDBCInput_2
-										.getInt(column_index_tJDBCInput_2);
+										.getLong(column_index_tJDBCInput_2);
 							} else {
 
 								throw new RuntimeException(
@@ -7307,9 +7307,9 @@ public class AggregationToDaily implements TalendJob {
 		final static byte[] commonByteArrayLock_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 		static byte[] commonByteArray_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 
-		public int history_id;
+		public long history_id;
 
-		public int getHistory_id() {
+		public long getHistory_id() {
 			return this.history_id;
 		}
 
@@ -7396,7 +7396,7 @@ public class AggregationToDaily implements TalendJob {
 
 					int length = 0;
 
-					this.history_id = dis.readInt();
+					this.history_id = dis.readLong();
 
 					this.history_datetime = readDate(dis);
 
@@ -7433,9 +7433,9 @@ public class AggregationToDaily implements TalendJob {
 		public void writeData(ObjectOutputStream dos) {
 			try {
 
-				// int
+				// long
 
-				dos.writeInt(this.history_id);
+				dos.writeLong(this.history_id);
 
 				// java.util.Date
 
@@ -7939,7 +7939,7 @@ public class AggregationToDaily implements TalendJob {
 							if (rs_tJDBCInput_3
 									.getObject(column_index_tJDBCInput_3) != null) {
 								row3.history_id = rs_tJDBCInput_3
-										.getInt(column_index_tJDBCInput_3);
+										.getLong(column_index_tJDBCInput_3);
 							} else {
 
 								throw new RuntimeException(
@@ -9888,9 +9888,9 @@ public class AggregationToDaily implements TalendJob {
 		final static byte[] commonByteArrayLock_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 		static byte[] commonByteArray_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 
-		public int history_id;
+		public long history_id;
 
-		public int getHistory_id() {
+		public long getHistory_id() {
 			return this.history_id;
 		}
 
@@ -10061,7 +10061,7 @@ public class AggregationToDaily implements TalendJob {
 
 					int length = 0;
 
-					this.history_id = dis.readInt();
+					this.history_id = dis.readLong();
 
 					this.history_datetime = readDate(dis);
 
@@ -10124,9 +10124,9 @@ public class AggregationToDaily implements TalendJob {
 		public void writeData(ObjectOutputStream dos) {
 			try {
 
-				// int
+				// long
 
-				dos.writeInt(this.history_id);
+				dos.writeLong(this.history_id);
 
 				// java.util.Date
 
@@ -10702,7 +10702,7 @@ public class AggregationToDaily implements TalendJob {
 							if (rs_tJDBCInput_4
 									.getObject(column_index_tJDBCInput_4) != null) {
 								row4.history_id = rs_tJDBCInput_4
-										.getInt(column_index_tJDBCInput_4);
+										.getLong(column_index_tJDBCInput_4);
 							} else {
 
 								throw new RuntimeException(
@@ -12392,9 +12392,9 @@ public class AggregationToDaily implements TalendJob {
 		final static byte[] commonByteArrayLock_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 		static byte[] commonByteArray_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 
-		public int history_id;
+		public long history_id;
 
-		public int getHistory_id() {
+		public long getHistory_id() {
 			return this.history_id;
 		}
 
@@ -12481,7 +12481,7 @@ public class AggregationToDaily implements TalendJob {
 
 					int length = 0;
 
-					this.history_id = dis.readInt();
+					this.history_id = dis.readLong();
 
 					this.history_datetime = readDate(dis);
 
@@ -12518,9 +12518,9 @@ public class AggregationToDaily implements TalendJob {
 		public void writeData(ObjectOutputStream dos) {
 			try {
 
-				// int
+				// long
 
-				dos.writeInt(this.history_id);
+				dos.writeLong(this.history_id);
 
 				// java.util.Date
 
@@ -13024,7 +13024,7 @@ public class AggregationToDaily implements TalendJob {
 							if (rs_tJDBCInput_5
 									.getObject(column_index_tJDBCInput_5) != null) {
 								row5.history_id = rs_tJDBCInput_5
-										.getInt(column_index_tJDBCInput_5);
+										.getLong(column_index_tJDBCInput_5);
 							} else {
 
 								throw new RuntimeException(
@@ -14313,9 +14313,9 @@ public class AggregationToDaily implements TalendJob {
 		final static byte[] commonByteArrayLock_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 		static byte[] commonByteArray_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 
-		public int history_id;
+		public long history_id;
 
-		public int getHistory_id() {
+		public long getHistory_id() {
 			return this.history_id;
 		}
 
@@ -14414,7 +14414,7 @@ public class AggregationToDaily implements TalendJob {
 
 					int length = 0;
 
-					this.history_id = dis.readInt();
+					this.history_id = dis.readLong();
 
 					this.history_datetime = readDate(dis);
 
@@ -14445,9 +14445,9 @@ public class AggregationToDaily implements TalendJob {
 		public void writeData(ObjectOutputStream dos) {
 			try {
 
-				// int
+				// long
 
-				dos.writeInt(this.history_id);
+				dos.writeLong(this.history_id);
 
 				// java.util.Date
 
@@ -14954,7 +14954,7 @@ public class AggregationToDaily implements TalendJob {
 							if (rs_tJDBCInput_11
 									.getObject(column_index_tJDBCInput_11) != null) {
 								row16.history_id = rs_tJDBCInput_11
-										.getInt(column_index_tJDBCInput_11);
+										.getLong(column_index_tJDBCInput_11);
 							} else {
 
 								throw new RuntimeException(
@@ -22153,9 +22153,9 @@ public class AggregationToDaily implements TalendJob {
 		final static byte[] commonByteArrayLock_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 		static byte[] commonByteArray_OVIRT_ENGINE_DWH_AggregationToDaily = new byte[0];
 
-		public int history_id;
+		public long history_id;
 
-		public int getHistory_id() {
+		public long getHistory_id() {
 			return this.history_id;
 		}
 
@@ -22338,7 +22338,7 @@ public class AggregationToDaily implements TalendJob {
 
 					int length = 0;
 
-					this.history_id = dis.readInt();
+					this.history_id = dis.readLong();
 
 					this.history_datetime = readDate(dis);
 
@@ -22410,9 +22410,9 @@ public class AggregationToDaily implements TalendJob {
 		public void writeData(ObjectOutputStream dos) {
 			try {
 
-				// int
+				// long
 
-				dos.writeInt(this.history_id);
+				dos.writeLong(this.history_id);
 
 				// java.util.Date
 
@@ -23025,7 +23025,7 @@ public class AggregationToDaily implements TalendJob {
 							if (rs_tJDBCInput_10
 									.getObject(column_index_tJDBCInput_10) != null) {
 								row13.history_id = rs_tJDBCInput_10
-										.getInt(column_index_tJDBCInput_10);
+										.getLong(column_index_tJDBCInput_10);
 							} else {
 
 								throw new RuntimeException(
@@ -25659,6 +25659,6 @@ public class AggregationToDaily implements TalendJob {
 	ResumeUtil resumeUtil = null;
 }
 /************************************************************************************************
- * 738823 characters generated by Talend Open Studio for Data Integration on the
- * May 13, 2014 4:05:01 PM IDT
+ * 738865 characters generated by Talend Open Studio for Data Integration on the
+ * August 10, 2014 3:08:10 PM IDT
  ************************************************************************************************/
